relocate core:android:jre into jvmShared source set
