+++ components/core/ktx/build.gradle.kts
@@ -7,17 +7,9 @@ android.namespace = "com.flipperdevices.core.ktx"
 
 commonDependencies {
     implementation(projects.components.core.log)
-
-    implementation(libs.appcompat)
-    implementation(libs.kotlin.coroutines)
+    implementation(projects.components.core.buildKonfig)
 
     implementation(libs.kotlin.coroutines)
-
-    // Compose
-    implementation(libs.compose.ui)
-    implementation(libs.compose.tooling)
-    implementation(libs.compose.foundation)
-    implementation(libs.compose.material)
 }
 
 commonTestDependencies {
