+++ components/core/ktx/src/jvmSharedMain/kotlin/com/flipperdevices/core/ktx/jre/StreamKtx.kt
